requirements.txt
Flask==2.3.2
google_api_python_client==2.86.0
httplib2==0.22.0
oauth2client==4.1.3
openai==0.27.4
protobuf==4.23.2
SpeechRecognition==3.10.0
